update script nova
update script nova

+++ OPS-U16/scripts/install/install_nova.sh
@@ -9,8 +9,10 @@ source $dir_path/../functions.sh
 
 ##  Init config path
 
-echocolor "Create DB for NOVA "
-cat << EOF | mysql -uroot -p$MYSQL_PASS
+if [ "$1" == "controller" ]; then
+		
+	echocolor "Create DB for NOVA"
+	cat << EOF | mysql -uroot -p$MYSQL_PASS
 CREATE DATABASE nova_api;
 CREATE DATABASE nova;
 GRANT ALL PRIVILEGES ON nova_api.* TO 'nova'@'localhost' IDENTIFIED BY '$NOVA_API_DBPASS';
@@ -19,25 +21,42 @@ GRANT ALL PRIVILEGES ON nova.* TO 'nova'@'localhost' IDENTIFIED BY '$NOVA_DBPASS
 GRANT ALL PRIVILEGES ON nova.* TO 'nova'@'%' IDENTIFIED BY '$NOVA_DBPASS';
 FLUSH PRIVILEGES;
 EOF
+else
+	echocolor "Khong phai node Controller, khong can cai DB"
+fi
 
-echocolor "Create user, endpoint for NOVA"
+if [ "$1" == "controller" ]; then
 
-openstack user create nova --domain default  --password $NOVA_PASS
-openstack role add --project service --user nova admin
-openstack service create --name nova --description "OpenStack Compute" compute
+	echocolor "Create user, endpoint for NOVA"
 
-openstack endpoint create --region RegionOne \
-    compute public http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
-openstack endpoint create --region RegionOne \
-    compute internal http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
-openstack endpoint create --region RegionOne \
-    compute admin http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
+	openstack user create nova --domain default  --password $NOVA_PASS
+	openstack role add --project service --user nova admin
+	openstack service create --name nova --description "OpenStack Compute" compute
+
+	openstack endpoint create --region RegionOne \
+		compute public http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
+	openstack endpoint create --region RegionOne \
+		compute internal http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
+	openstack endpoint create --region RegionOne \
+		   compute admin http://$CTL_MGNT_IP:8774/v2.1/%\(tenant_id\)s
+else
+	echocolor "Khong phai node Controller, khong can tao endpoint"
+fi
 
 echocolor "Install NOVA in $CTL_MGNT_IP"
-sleep 5
-apt-get -y install nova-api nova-cert \
-    nova-conductor nova-consoleauth \
-    nova-novncproxy nova-scheduler
+sleep 3
+
+if [ "$1" == "controller" ]; then
+
+	apt-get -y install nova-api nova-cert nova-conductor nova-consoleauth \
+	    	nova-novncproxy nova-scheduler
+
+elif [ "$1" == "compute1" ] || [ "$1" == "compute2" ] ; then
+	 apt-get -y install nova-compute
+
+else 
+	 echocolor "Khong phai node COMPUE"
+fi
 
 ######## Backup configurations for NOVA ##########"
 sleep 7
@@ -56,11 +75,22 @@ ops_edit $nova_ctl DEFAULT enabled_apis osapi_compute,metadata
 ops_edit $nova_ctl DEFAULT rpc_backend rabbit
 ops_edit $nova_ctl DEFAULT auth_strategy keystone
 ops_edit $nova_ctl DEFAULT rootwrap_config /etc/nova/rootwrap.conf
-ops_edit $nova_ctl DEFAULT my_ip $CTL_MGNT_IP
 ops_edit $nova_ctl DEFAULT use_neutron True
 ops_edit $nova_ctl DEFAULT \
     firewall_driver nova.virt.firewall.NoopFirewallDriver
 ops_edit $nova_ctl DEFAULT transport_url  rabbit://openstack:$RABBIT_PASS@$CTL_MGNT_IP
+
+if [ "$1" == "controller" ]; then
+	ops_edit $nova_ctl DEFAULT my_ip $CTL_MGNT_IP
+
+elif [ "$1" == "compute1" ]; then
+	ops_edit $nova_ctl DEFAULT my_ip $COM1_MGNT_IP
+
+elif [ "$1" == "compute2" ]; then
+	ops_edit $nova_ctl DEFAULT my_ip $COM2_MGNT_IP
+else
+	echocolor "Khong phai node cai nova"
+fi
 
 ## [api_database] section
 ops_edit $nova_ctl api_database \
@@ -88,8 +118,19 @@ ops_edit $nova_ctl keystone_authtoken username nova
 ops_edit $nova_ctl keystone_authtoken password $NOVA_PASS
 
 ## [vnc] section
-ops_edit $nova_ctl vnc vncserver_listen \$my_ip
-ops_edit $nova_ctl vnc vncserver_proxyclient_address \$my_ip
+if [ "$1" == "controller" ]; then
+	ops_edit $nova_ctl vnc vncserver_listen \$my_ip
+	ops_edit $nova_ctl vnc vncserver_proxyclient_address \$my_ip
+
+elif [ "$1" == "compute1" ] || [ "$1" == "compute2" ] ; then
+	ops_edit $nova_ctl vnc enabled  true
+	ops_edit $nova_ctl vnc vncserver_listen 0.0.0.0
+	ops_edit $nova_ctl vnc vncserver_proxyclient_address \$my_ip
+	ops_edit $nova_ctl vnc novncproxy_base_url http://$CTL_MGNT_IP:6080/vnc_auto.htm
+else
+	echo "Khong can cai VNC"
+fi
+	
 
 ## [glance] section
 ops_edit $nova_ctl glance api_servers http://$CTL_MGNT_IP:9292
@@ -110,35 +151,43 @@ ops_edit $nova_ctl neutron password $NEUTRON_PASS
 ops_edit $nova_ctl neutron service_metadata_proxy True
 ops_edit $nova_ctl neutron metadata_proxy_shared_secret $METADATA_SECRET
 
-## [cinder] Section
-ops_edit $nova_ctl cinder os_region_name RegionOne
+if [ "$1" == "controller" ]; then 
+	
+	## [cinder] Section
+	ops_edit $nova_ctl cinder os_region_name RegionOne
+	echocolor "Remove Nova default db "
+	sleep 5
+	rm /var/lib/nova/nova.sqlite
 
-echocolor "Remove Nova default db "
-sleep 5
-rm /var/lib/nova/nova.sqlite
+	echocolor "Syncing Nova DB"
+	sleep 5
+	su -s /bin/sh -c "nova-manage api_db sync" nova
+	su -s /bin/sh -c "nova-manage db sync" nova
 
-echocolor "Syncing Nova DB"
-sleep 5
-su -s /bin/sh -c "nova-manage api_db sync" nova
-su -s /bin/sh -c "nova-manage db sync" nova
 
-echocolor "Restarting NOVA "
-sleep 7
-service nova-api restart
-service nova-cert restart
-service nova-consoleauth restart
-service nova-scheduler restart
-service nova-conductor restart
-service nova-novncproxy restart
+	echocolor "Restarting NOVA "
+	sleep 7
+	service nova-api restart
+	service nova-cert restart
+	service nova-consoleauth restart
+	service nova-scheduler restart
+	service nova-conductor restart
+	service nova-novncproxy restart
 
-sleep 7
-echocolor "Restarting NOVA"
-service nova-api restart
-service nova-cert restart
-service nova-consoleauth restart
-service nova-scheduler restart
-service nova-conductor restart
-service nova-novncproxy restart
+	sleep 7
+	echocolor "Restarting NOVA"
+	service nova-api restart
+	service nova-cert restart
+	service nova-consoleauth restart
+	service nova-scheduler restart
+	service nova-conductor restart
+	service nova-novncproxy restart
 
-echocolor "Testing NOVA service"
-openstack compute service list
+	echocolor "Testing NOVA service"
+	openstack compute service list
+
+elif [ "$1" == "compute1" ] || [ "$1" == "compute2" ] ; then
+	service nova-compute restart
+
+else
+	echocolor "Khong phai NOVA - CTL"
